annotation_service/requirements.txt
kafka-python
boto3
requests
python-dotenv
